Fix global analysis decode and navigation for US/crypto detail screens
Week52Range, BollingerBands, and VolumeAnalysis still used Int fields,
so fractional US/crypto values (e.g. NVDA's 52-week high) failed the
whole CandleAnalysis decode and the technical tab showed dashes. The
global market rows also switch to destination-style NavigationLinks
since value-based links weren't resolving inside the pushed stack.

+++ AlphaTradingIOS/AlphaTrading/Models/StockAnalysis.swift
@@ -271,21 +271,21 @@ struct MADistance: Decodable {
 }
 
 struct BollingerBands: Decodable {
-    let upper: Int?
-    let mid: Int?
-    let lower: Int?
+    let upper: Double?
+    let mid: Double?
+    let lower: Double?
     let bandwidth: Double?
     let position: Double?
 }
 
 struct Week52Range: Decodable {
-    let high: Int?
-    let low: Int?
+    let high: Double?
+    let low: Double?
     let position: Double?
 }
 
 struct VolumeAnalysis: Decodable {
-    let latest: Int?
+    let latest: Double?
     let avg20: Double?
     let ratio: Double?
 }
